+++ components/src/lib/constants/modal.ts
@@ -1,11 +1,18 @@
+import { CSSProperties } from 'styled-components';
+
 export const BASIC_BOTTOM_MODAL_ANIMATION_DURATION = 600;
 export const NOW = 0;
 export const BASIC_TOAST_MODAL_ANIMATION_DURATION = 3000;
 export const BASIC_TOAST_DURATION = 6000;
-export const CONTENTS_CLASS_NAME = 'modal-contents';
 export const BASIC_PADDING = '2rem 1.5rem';
 export const BASIC_BORDER_RADIUS = '0.625rem';
 export const BASIC_BACKGROUND_COLOR = {
   modal: '#ffff',
   backdrop: '#5959599b',
+};
+export const BASIC_BUTTON_STYLE: CSSProperties = {
+  padding: '0.625rem 0.875rem',
+  backgroundColor: '#ffff',
+  color: 'black',
+  fontSize: '1rem',
 };
